// admin/src/api/generated/models/HistoryPoint.ts
export interface HistoryPoint {
    /**
     *
     * @type {number}
     * @memberof HistoryPoint
     */
    cpuUsageRatio?: number | null;
    /**
     *
     * @type {number}
     * @memberof HistoryPoint
     */
    diskBusyRatio?: number | null;
    /**
     *
     * @type {number}
     * @memberof HistoryPoint
     */
    diskReadBytesPerSecond?: number | null;
    /**
     *
     * @type {number}
     * @memberof HistoryPoint
     */
    diskWriteBytesPerSecond?: number | null;
    /**
     *
     * @type {number}
     * @memberof HistoryPoint
     */
    memoryUsedBytes?: number | null;
    /**
     *
     * @type {number}
     * @memberof HistoryPoint
     */
    networkReceiveBytesPerSecond?: number | null;
    /**
     *
     * @type {number}
     * @memberof HistoryPoint
     */
    networkTransmitBytesPerSecond?: number | null;
    /**
     *
     * @type {string}
     * @memberof HistoryPoint
     */
    receivedAt: string;
    /**
     *
     * @type {number}
     * @memberof HistoryPoint
     */
    workRootUsedBytes?: number | null;
}